refactor envoy filters
